+++ src/Catalyst.jl
@@ -112,7 +112,7 @@ export hc_steady_states
 
 # spatial reactions
 include("spatial_reaction_systems/spatial_reactions.jl")
-export TransportReaction, transport_reactions, @transport_reaction
+export TransportReaction, TransportReactions, @transport_reaction
 export isedgeparameter
 
 # lattice reaction systems