+++ HISTORY.md
@@ -2,7 +2,7 @@
 
 ## Catalyst unreleased (master branch)
 - Simulation of spatial ODEs now supported. For full details, please see https://github.com/SciML/Catalyst.jl/pull/644 and upcoming documentation.
-- LatticeReactionSystem structure represents a spatiral reaction network:
+- LatticeReactionSystem structure represents a spatial reaction network:
   ```julia
   rn = @reaction_network begin
       (p,d), 0 <--> X
@@ -10,12 +10,20 @@
   tr = @transport_reaction D X
   lattice = Graphs.grid([5, 5])
   lrs = LatticeReactionSystem(rn, [tr], lattice)
+- Here, if a `u0` or `p` vector is given with scalar values:
+  ```julia
+  u0 = [:X => 1.0]
+  p = [:p => 1.0, :d => 0.5, :D => 0.1]
   ```
-
+  this value will be used across the entire system. If their values are instead vectors, different values are used across the spatial system. Here
+  ```julia
+  u0 = [:X => [1.0, 0.0, 0.0, ...]]
+  ```
+  X's value will be `1.0` in the first vertex, but `0.0` in the remaining one (the system have 25 vertexes in total). SInce th parameters `p` and `d` are part of the non-spatial reaction network, their values are tied to vertexes. However, if the `D` parameter (which governs diffusion between vertexes) is given several values, these will instead correspond to the specific edges (and transportation along those edges.)
 
 
 ## Catalyst 13.5
-- Added a CatalystHomotopyContinuationExtension extension, which exports the `hc_steady_state` function if HomotopyContinuation is exported. `hc_steady_state` finds the steady states of a reactin system using the homotopy continuation method. This feature is only available for julia versions 1.9+. Example: 
+- Added a CatalystHomotopyContinuationExtension extension, which exports the `hc_steady_state` function if HomotopyContinuation is exported. `hc_steady_state` finds the steady states of a reaction system using the homotopy continuation method. This feature is only available for julia versions 1.9+. Example: 
 ```julia
 wilhelm_2009_model = @reaction_network begin
     k1, Y --> 2X